Test case generator updated

+++ Test Case Generator/gen.py
@@ -1,21 +1,26 @@
 import os,time
-os.system("rm test.zip")
+
+if(os.path.isfile('test.zip')):
+    os.system("rm test.zip")
+
 os.system("g++ test_case.cpp")
-number_of_cases=10
+
+number_of_cases=5
+x=5
 for i in range(number_of_cases):
-    if i+10<10:
-        os.system("./a.out <temp> input/input0"+str(i+10)+".txt")
+    if i+x<10:
+        os.system("./a.out > input/input0"+str(i+x)+".txt")
     else:
-        os.system("./a.out <temp> input/input"+str(i+10)+".txt")
+        os.system("./a.out > input/input"+str(i+x)+".txt")
     time.sleep(1)
 
 os.system("g++ answer.cpp")
 
 for i in range(number_of_cases):
-    if i+10<10:
-        os.system("./a.out <input/input0"+str(i+10)+".txt> output/output0"+str(i+10)+".txt")
+    if i+x<10:
+        os.system("./a.out < input/input0"+str(i+x)+".txt > output/output0"+str(i+x)+".txt")
     else:
-        os.system("./a.out <input/input"+str(i+10)+".txt> output/output"+str(i+10)+".txt")
+        os.system("./a.out < input/input"+str(i+x)+".txt > output/output"+str(i+x)+".txt")
 
 s="zip -r test input output"
 os.system(s)
